Refactor package locations
