+++ src/main/java/com/kw/evcs/domain/entity/User.java
@@ -1,4 +1,4 @@
-package com.kw.evcs.entity;
+package com.kw.evcs.domain.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
