+++ src/main/java/com/kw/evcs/domain/repository/AgencyRepository.java
@@ -1,6 +1,6 @@
-package com.kw.evcs.repository;
+package com.kw.evcs.domain.repository;
 
-import com.kw.evcs.entity.Agency;
+import com.kw.evcs.domain.entity.Agency;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AgencyRepository extends JpaRepository<Agency, Long> {
