+++ src/main/java/com/kw/evcs/domain/repository/ChargerRepository.java
@@ -1,6 +1,6 @@
-package com.kw.evcs.repository;
+package com.kw.evcs.domain.repository;
 
-import com.kw.evcs.entity.Charger;
+import com.kw.evcs.domain.entity.Charger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChargerRepository extends JpaRepository<Charger, Long> {
