+++ src/main/java/com/kw/evcs/domain/repository/ChargingStationRepository.java
@@ -1,6 +1,6 @@
-package com.kw.evcs.repository;
+package com.kw.evcs.domain.repository;
 
-import com.kw.evcs.entity.ChargingStation;
+import com.kw.evcs.domain.entity.ChargingStation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChargingStationRepository extends JpaRepository<ChargingStation, Long> {
