+++ src/main/java/com/kw/evcs/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.kw.evcs.repository;
+package com.kw.evcs.domain.repository;
 
-import com.kw.evcs.entity.User;
+import com.kw.evcs.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
